Try using sqlite in production
Switch over to using sqlite in production and add a step during the
container build to pack the initial image data in the container. This
speeds up the startup time and also makes it work better in air-gapped
environments.

We're able to remove the threading/scheduling from the main module by
using this method.

Signed-off-by: Major Hayden <major@redhat.com>